8018 Xen HVM VM with 32 VCPUs hangs on boot
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: Igor Kozhukhov <igor@dilos.org>
Reviewed by: Robert Mustacchi <rm@joyent.com>

The `pending_word` code is a bitmap which has a bit set for each event
channel that has pending data. The code processes one bit at a time, and
turns off bits as they are processed. The problem is on this line:

    pending_word = pending_word & ~(1 << j);

The integer constant `1` defaults to an `int` type so if `j > 31`, the
result of `1 << j` will be `0`. This means that we never turn off the
higher order bits and get stuck in an infinite loop. The fix is to
simply declare the integer constant as an `unsigned long`. I scanned
through the rest of the Xen code for any other bit shifts to see if
there might be other instances of this bug but did not find any. I was
a bit curious why this bug didn't show up when we used VMs with smaller
VCPU counts. With `kmdb` I was able to verify that the event channel
identifiers used for smaller VCPU configurations did not exceed 32:

    # (rdi is the event channel identifier)
    %rdi = 0x0000000000000019
    %rdi = 0x000000000000001b
    %rdi = 0x000000000000001c
    %rdi = 0x000000000000001d
    %rdi = 0x000000000000001e

Once we enabled 32 VCPUS we saw larger event channel identifiers:

    %rdi = 0x0000000000000021
    %rdi = 0x0000000000000023
    %rdi = 0x0000000000000024
    %rdi = 0x0000000000000025
    %rdi = 0x0000000000000026

Upstream bugs: 34224