8019 Some PV devices should not be configured in HVM mode
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: Basil Crow <basil.crow@delphix.com>

When booting a Xen VM in HVM mode, we can see the `/var/adm/messages`
file is being spammed with the following messages:

    May 16 07:01:06 delphix genunix: [ID 408114 kern.info] /xpvd/xencons@0 (xencons0) offline
    May 16 07:01:06 delphix genunix: [ID 408114 kern.info] /xpvd/xenbus@0 (xenbus0) offline
    May 16 07:01:06 delphix genunix: [ID 408114 kern.info] /xpvd/domcaps@0 (domcaps0) offline
    May 16 07:01:06 delphix genunix: [ID 408114 kern.info] /xpvd/balloon@0 (balloon0) offline

This is happening because we are trying to configure Xen PV devices that
aren't available while in HVM mode. The configuration of these devices
fails, which causes us to to send a sysevent announcing that the event is
offline. This sysevent is being processed by devfsadm and causes it to
update it's device topology snapshot. Updating the device topology
causes xpvd to try and reconfigure its devices so we end up in an
infinite loop.

The devices that we are failing to configure are not available in HVM
mode and it does not make sense to try and create such devices.

Upstream bugs: 34424