7777 Expose xdf minor nodes when in PV-HVM mode
Reviewed by: Jeremy Jones jeremy@delphix.com
Reviewed by: Basil Crow basil.crow@delphix.com

When in HVM mode the minor nodes of the xdf devices are hidden since the
xdf device is layered below an IDE disk device. In PV-HVM mode we have
disabled IDE devices and need to expose the xdf minor nodes in order for
devices to be visible in the device tree.

Upstream bugs: 34508