7496 cmp_ds_cont has never worked
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>

The "cmp_ds_cont" function in the zfs-test suite ends with a call to
"echo". As a result, any consumers of this function interpreting its
return code to mean "success" or "failure" would have been flawed
(i.e. if it was used in conjunction with "log_must"), as the return
code of the function would have been the return code of the "echo"
command; i.e. it would almost always report success.

This patch modifies this function so that its return code can be
properly used to convey "success" or "failure", and then its usage
was removed from a few tests where it shouldn't have been used to
begin with.

Upstream Bugs: QA-6101