7543 ::spa -h should show free space histogram for whole pool
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

This patch modifies the mdb dcmd "::spa", such that when the "-h" option
is passed in, a histogram of free space for the entire pool will be
shown. Additionally, "-h" will no longer implies printing the device
listing, although other flags for this dcmd sill do.

Here's an example of the new output for this dcmd:

    > ::spa -h
    ADDR                 STATE NAME
    fffffeebc1c3d000    ACTIVE dcenter
          9: 5045975 *****
         10: 9172273 ********
         11: 16709142 ***************
         12: 29128030 *************************
         13: 38901639 *********************************
         14: 47459983 ****************************************
         15: 43807301 *************************************
         16: 22494772 *******************
         17: 4909315 *****
         18: 292411 *
         19:   9724 *
         20:     40 *

Upstream bugs: DLPX-39344