7604 if volblocksize property is the default, it displays as "-" rather than 8K
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>
Reviewed by: Paul Dagnelie <pcd@delphix.com>
Reviewed by: John Kennedy <john.kennedy@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

If a zvol has the default setting for the "volblocksize" property, it is
8KB.  However, it is displayed as "-" (not present), rather than "8K".

The problem was introduced by:

    7571 non-present readonly numeric ZFS props do not have default value

which changed changed get_numeric_property() to indicate that readonly
default properties are not present.  However, zfs_prop_readonly() returns
TRUE for both readonly and set-once properties (e.g. volblocksize).

Amusingly, that commit essentially reverted:

    6900484 default volblocksize is no longer being reported correctly

from November 2009.  However, that change was not correct either; the
correct solution is to only do this check for "truly readonly" (i.e. not
setonce) properties.

Upstream bugs: DLPX-45214