7741 get_zfs_dataset() takes an unreasonable amount of time
Reviewed by: Steve Gonczi <steve.gonczi@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>

get_zfs_dataset() currently takes an unreasonable amount of time in the
process of sharing filesystems. It does multiple lookups of zfs
properties per filesystem on the machine, per filesystem that needs to
be shared (yes, quadratic in most cases).

The naive solution implemented in this change, is to make an optimistic
assumption that the mountpoint for a zfs filesystem is within a few
leading slashes of its name. So we cut off those leading slashes and
then check to see if such a filesystem exists. Only if the filesystem
has a custom mountpoint does this not work, so we turn a quadratic
operation into a constant time one (most of the time).

To verify the improvement, 10,000 zfs filesystems were created on a VM,
and the following command was used to share them all:

    # sharemgr start zfs

Without this change, this operation took about 6.5 minutes; and with
this change, it took about 6.8 seconds.

Upstream bugs: DLPX-45572