Port hole birth tests from zfsonlinux #5675 to openzfs
This is intended to be a follow on for the openzfs PR #273, and pulls in
the new test cases added in the zfsonlinux PR #5675.

I haven't built or tested this patch yet, so there's a good chance
it's broken, but I wanted to open a PR so I can run this through the
automated system.

I generated this patch by configuring git to allow cherry-picking the
zfsonlinx commit directly; i.e.

    $ git config merge.renameLimit 999999
    $ git cherry-pick -X ours 29031d3660991b446c6fd7dfbf49c7c58ee211ae

and then I had to manually fix up a few files due to differences in the
build system for two repositories (e.g. remove Makefile.am, etc).

Here's a link to the zfsonlinx PR: https://github.com/zfsonlinux/zfs/pull/5675