7955 libshare needs to initialize only those datasets being modified by the consumer
Reviewed by: Steve Gonczi <steve.gonczi@delphix.com>
Reviewed by: Sebastien Roy <sebastien.roy@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>

Libshare currently initializes all available filesystems when doing any
libshare operation. This requires iterating through all the filesystem
multiple times, which is a huge performance problem for sharing and
unsharing operations.

This patch modifies the initialization of libshare in two more modes
beyond what had happened before. In one, a pathname is passed in, and
that one path is intiialized if possible. In another mode of operation,
an array of zfs dataset handles is passed into libshare to be
initialized. In any case, these initialization operations should speed
up most uses of libshare.

The following tests were run on a system with ~8300 ZFS filesystems on
it, using time(1) to measure the performance of each command without
this patch:

    $ sudo zfs create -o sharenfs=on rpool/test
    $ time sudo zfs destroy rpool/test
    real    0m28.957s
    user    0m7.494s
    sys     0m20.969s

    $ time sudo zfs create -o sharenfs=on rpool/test
    real    0m29.216s
    user    0m7.517s
    sys     0m20.579s

    $ time sudo zfs destroy rpool/test
    real    0m28.527s
    user    0m7.435s
    sys     0m20.688s

And then same commands were run on the same system, but using a version
of libshare with this patch applied:

    $ sudo zfs create -o sharenfs=on rpool/test
    $ time sudo zfs destroy rpool/test
    real    0m0.624s
    user    0m0.067s
    sys     0m0.255s

    $ time sudo zfs create -o sharenfs=on rpool/test
    real    0m0.797s
    user    0m0.020s
    sys     0m0.183s

    $ time sudo zfs destroy rpool/test
    real    0m0.627s
    user    0m0.072s
    sys     0m0.270s

That's about a 97% decrease in "wall clock" latency for these commands.

Upstream bugs: DLPX-47705, DLPX-47160, DLPX-48226, DLPX-50300