8024 mdb_ctf_vread() needn't be so strict about unions
Reviewed by: Matthew Ahrens <mahrens@delphix.com>
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>

This modifies mdb's vread_helper function such that not all target union
members need to be present in the mdb type. If there is only a single
union member in the mdb type, its actual type does not need to match
with its target's type. On the other hand, if more than one union
members are specified in the mdb type, their types must match with the
types of their relevant union members of the target union.

Upstream bugs: DLPX-23010