7762 avoid division by zero in property_alias_001_pos
Reviewed by: John Kennedy <john.kennedy@delphix.com>
Reviewed by: Matt Ahrens <mahrens@delphix.com>

The `property_alias_001_pos` test case of the zfstest suite attempts
to perform division using `$RANDOM` as the divisor. Thus, if `$RANDOM`
happens to evaluate to zero, the test will fail. We need to update the
test to avoid this potential failure mode, which can be done by simply
adding one to the value of `$RANDOM` prior to performing the division.